Move properties package to be a child package of plugins
