+++ squidb-processor/src/com/yahoo/squidb/processor/plugins/properties/factory/FieldReferencePropertyGeneratorFactory.java
@@ -3,14 +3,14 @@
  * Copyrights licensed under the Apache 2.0 License.
  * See the accompanying LICENSE file for terms.
  */
-package com.yahoo.squidb.processor.properties.factory;
+package com.yahoo.squidb.processor.plugins.properties.factory;
 
 import com.yahoo.aptutils.model.DeclaredTypeName;
 import com.yahoo.aptutils.utils.AptUtils;
 import com.yahoo.squidb.processor.TypeConstants;
 import com.yahoo.squidb.processor.plugins.Plugin;
-import com.yahoo.squidb.processor.properties.generators.PropertyGenerator;
-import com.yahoo.squidb.processor.properties.generators.ViewPropertyGenerator;
+import com.yahoo.squidb.processor.plugins.properties.generators.PropertyGenerator;
+import com.yahoo.squidb.processor.plugins.properties.generators.ViewPropertyGenerator;
 
 import javax.lang.model.element.TypeElement;
 import javax.lang.model.element.VariableElement;
